FNN
Created a model for feed forward neural networks that estimates market value of players using a Leaky ReLU activation function, Adam optimizer, and MSE loss function.
This is, however, and early iteration of the final product. wee see that small changes in features result in large changes in target value.